main() function refactored

+++ accountant.py
@@ -30,20 +30,14 @@ import sys
 import os.path
 
 
-AVAILABLE_COMMANDS = ("saldo", "zakup", "sprzedaż", "konto", "magazyn", "przegląd")
+AVAILABLE_COMMANDS = ("saldo", "zakup", "sprzedaż", "konto", "magazyn", 
+"przegląd")
 commands = sys.argv
 
 def main():
     
-    # Check if konto.txt file exists
-    if os.path.isfile("konto.txt"):
-        print()
-        account()
- 
-    # konto.txt file doesn't exist - create a file and initiate account ballance info
-    else:
-        with open("konto.txt", "w") as account_file:
-            account_file.write("saldo: 0")
+    # Check if database already exists - otherwise create file konto.txt
+    start_database()
     
     # Execute code based on user command
     with open("magazyn.txt", "a"):
@@ -60,188 +54,115 @@ e) python accountant.py magazyn <str identyfikator produktu 1> <str identyfikato
 f) python accountant.py przegląd""")         
 
         if commands[1] not in AVAILABLE_COMMANDS:    
-            sys.exit("Podaj poprawną komendę")
-           
-        if commands[1] == "zakup":
-            buy()
-            # Add current command to log.txt
-            save_to_log()
-            account()
-          
-        if commands[1] == "saldo":
+            sys.exit("Podaj poprawną komendę")   
+        elif commands[1] == "zakup":
+            buy() 
+            account()  
+        elif commands[1] == "saldo":
             saldo()
-            # Add current command to log.txt
-            save_to_log()
-            account()
-                         
-        if commands[1] == "sprzedaż":    
+            account()                
+        elif commands[1] == "sprzedaż":    
             sell()
-            # Add current command to log.txt
-            save_to_log()
+            account()           
+        elif commands[1] == "konto":    
             account()
-                    
-        if commands[1] == "konto":    
-            account()
+        elif commands[1] == "magazyn":
+            view_storage()   
+        elif commands[1] == "przegląd":     
             # Add current command to log.txt
-            save_to_log()
-        
-        if commands[1] == "magazyn":
-            view_storage()
-            # Add current command to log.txt
-            save_to_log()
-            
-        if commands[1] == "przegląd":     
-            # Add current command to log.txt
-            save_to_log()
             view_log()
-               
+        save_to_log()       
     
 "View current account"
 def account():
-
     # Add current command to log.txt
-    with open("konto.txt", "r") as account_file:
-            
+    with open("konto.txt", "r") as account_file:        
         for line in account_file:    
             if line.startswith("saldo:"):    
                 line = line.strip().split()
                 print(f"Saldo konta: {line[1]}")
-       
-            
+                   
 "Buy products"
 def buy():
-
-    try:    
-        product = str(commands[2])
-    
-    except ValueError or IndexError:    
-        sys.exit("Podaj indentyfikator produktu")
-    
-    try:    
-        price = int(commands[3])
-    
-    except ValueError or IndexError:    
-        sys.exit("Podaj cenę za sztukę w groszach")
-        
-    
-    try:    
-        purchase_quantity = int(commands[4])
-    
-    except ValueError or IndexError:    
-        sys.exit("Podaj ilość zakupionego produktu")
-          
-    
-    if price < 1 or purchase_quantity < 1:    
-        sys.exit("Cena ani ilość nie mogą być mniejsze niż 1") 
-        
-    
-    # Total purchase value
-    total = price * purchase_quantity
-  
+    # Check if user's commands are correct   
+    product, price, quantity, total = validate_user_inputs()  
     # Check available funds 
     current_account = 0
-
-    
-    with open("konto.txt", "r") as account_file:
-        
-        for line in account_file:
-            
+   
+    with open("konto.txt", "r") as account_file:    
+        for line in account_file:    
             if line.startswith("saldo:"):    
                 line = line.strip().split()
-                current_account = int(line[1])
-                
+                current_account = int(line[1])    
                 if current_account < total:        
-                    sys.exit("Brak wystarczającej ilości środków na koncie")
-                        
+                    sys.exit("Brak wystarczającej ilości środków na koncie")        
                 else:    
                     current_account -= total
 
     with open("konto.txt", "w") as account_file:
         account_file.write(f"saldo: {current_account}")
     
-    # Print magazyn.txt file contents to a string temp_file replacing the changed line with new_line
+    # Print magazyn.txt file contents to a string temp_file replacing the 
+    # changed line with new_line
     temp_file = ""
     product_in_file = False
     
-    with open("magazyn.txt", "r") as store_file:
-        
+    with open("magazyn.txt", "r") as store_file:    
         # Check if product is already in the store and fetch its quantity
         for line in store_file:
-
             if line.startswith(product):     
                 split_line = line.strip().split()
                 store_quantity = int(split_line[1])
-                new_store_quantity = store_quantity + purchase_quantity
+                new_store_quantity = store_quantity + quantity
                 new_line = f"{product} {new_store_quantity}\n"
                 temp_file += new_line
                 product_in_file = True
-            
             else:
                 temp_file += f"{line}\n"
-        
         if not product_in_file:    
-            temp_file += f"{product} {purchase_quantity}\n"
+            temp_file += f"{product} {quantity}\n"
            
     # Print temp_file contents to file
-    with open("magazyn.txt", "w") as store_file:
-        
+    with open("magazyn.txt", "w") as store_file:    
         store_file.write(temp_file)
        
-    with open("tranzakcje.txt", "a") as transactions_file:
-        
+    with open("tranzakcje.txt", "a") as transactions_file:    
         transaction_str = " ". join(commands[1: ])
         transactions_file.write(transaction_str + "\n")
-    
     print(" ".join(commands))
  
-
 "View storage"
-def view_storage():
-    
+def view_storage():    
     if len(commands) < 3:    
-        sys.exit("Podaj id produktu")
-        
-    
+        sys.exit("Podaj id produktu")          
     # Read magazyn.txt file and print requested products' data
-    with open("magazyn.txt", "r") as store_file:
-        #print(store_file.read())
-            
-        for line in store_file:
-            
+    with open("magazyn.txt", "r") as store_file:       
+        for line in store_file:   
             line = line.strip()
-            
-            for command in commands[2:]:
-                
+            for command in commands[2:]:    
                 if line.startswith(command):    
                     searched_line = line.split()
                     print(": ".join(searched_line))
 
-
 "Register a new transaction"
-def saldo():
-    
+def saldo():  
     try:    
         value = int(commands[2])
-    
     except ValueError or IndexError:    
         sys.exit("Podaj kwotę w groszach")
         
     try:    
         comment = str(commands[3])
-    
     except ValueError or IndexError:    
         sys.exit("Podaj komantarz do zmiany salda")
 
-    with open("tranzakcje.txt", "a") as transactions_file:
-        
+    with open("tranzakcje.txt", "a") as transactions_file:    
         transactions_file.write(f"{commands[1]} {value} {comment}\n")
     
     current_account = 0
     
-    with open("konto.txt", "r") as account_file:
-        
-        for line in account_file:
-            
+    with open("konto.txt", "r") as account_file:   
+        for line in account_file:    
             if line.startswith("saldo:"):    
                 line = line.strip().split()
                 current_account += int(line[1])
@@ -249,50 +170,21 @@ def saldo():
     current_account += value
 
     with open("konto.txt", "w") as account_file:
-
         account_file.write(f"saldo: {current_account}")
-   
     print(" ".join(commands))
 
-
 "Save current command to log.txt - helper function"
-def save_to_log():
-    
+def save_to_log():  
     with open("log.txt", "a") as log_file:
-
         commands_str = " ".join(commands)
         log_file.write(commands_str + "\n")
 
-
 "Sell products"
 def sell():
-
-    try:    
-        product = str(commands[2])
-    
-    except ValueError or IndexError:    
-        print("Podaj indentyfikator produktu")
-    
-    try:    
-        price = int(commands[3])
-    
-    except ValueError or IndexError:    
-        sys.exit("Podaj cenę za sztukę w groszach")
-          
-    try:    
-        sell_quantity = int(commands[4])
-    
-    except ValueError or IndexError:    
-        sys.exit("Podaj ilość sprzedanego produktu")
-        
-    total = price * sell_quantity
-
-    if price < 1 or sell_quantity < 1:    
-        sys.exit("Cena ani ilość nie mogą być mniejsze niż 1") 
-
+    # Check if user's commands are correct
+    product, price, quantity, total = validate_user_inputs()
     # Check availability of the product
     new_store_quantity = 0
-
     # Print file contents to string and replace the changed line with new_line
     new_line = f"{product} {new_store_quantity}\n"
     temp_file = ""
@@ -304,29 +196,23 @@ def sell():
             # Update line contents with the requested amount of the product is available in the store
             if line.startswith(product):    
                 line = line.strip().split()
-                current_store_quantity = int(line[1])
-                
-                if current_store_quantity >= sell_quantity:    
-                    new_store_quantity = current_store_quantity - sell_quantity
-                    temp_file += new_line
-                
+                current_store_quantity = int(line[1])    
+                if current_store_quantity >= quantity:    
+                    new_store_quantity = current_store_quantity - quantity
+                    temp_file += new_line   
                 else:    
-                    sys.exit("Brak wystarczającej ilości produktu w magazynie")
-              
+                    sys.exit("Brak wystarczającej ilości produktu w magazynie")   
             else:
                 temp_file += f"{line}\n"
 
-
     # Write temp_file contents to file
-    with open("magazyn.txt", "w") as store_file:
-        
+    with open("magazyn.txt", "w") as store_file:   
         store_file.write(temp_file)
        
     # Find current account balance
     current_account = 0
     
-    with open("konto.txt", "r") as account_file:
-        
+    with open("konto.txt", "r") as account_file:    
         for line in account_file:
             
             if line.startswith("saldo:"):    
@@ -336,29 +222,57 @@ def sell():
     current_account += total
 
     # Write current account balance to konto.txt file
-    with open("konto.txt", "w") as account_file:
-        
+    with open("konto.txt", "w") as account_file:    
         account_file.write(f"saldo: {current_account}")
    
     # Write current transaction to tranzakcje.txt file
-    with open("tranzakcje.txt", "a") as transactions_file:
-        
+    with open("tranzakcje.txt", "a") as transactions_file:    
         transaction_str = " ". join(commands[1: ])
         transactions_file.write(transaction_str + "\n")
     
     print(" ".join(commands))
 
-
 "View history of commands"
 def view_log():
     
     # Print all historical commands
-    with open("log.txt", "r") as log_file:
-        
+    with open("log.txt", "r") as log_file:    
         print(log_file.read(), sep="/n")
-    
     sys.exit()
 
+"Helper function - check if database exists"
+def start_database():
+    # Check if konto.txt file exists
+    if os.path.isfile("konto.txt"):
+        print()
+        account()
+    # konto.txt file doesn't exist - create a file and initiate account ballance info
+    else:
+        with open("konto.txt", "w") as account_file:
+            account_file.write("saldo: 0")
+
+"Helper function - check if user entered correct commands"
+def validate_user_inputs():
+    try:    
+        product = str(commands[2]) 
+    except ValueError or IndexError:    
+        print("Podaj indentyfikator produktu")
+    
+    try:    
+        price = int(commands[3]) 
+    except ValueError or IndexError:    
+        sys.exit("Podaj cenę za sztukę w groszach")
+          
+    try:    
+        quantity = int(commands[4]) 
+    except ValueError or IndexError:    
+        sys.exit("Podaj ilość produktu")
+        
+    total = price * quantity
+
+    if price < 1 or quantity < 1:    
+        sys.exit("Cena ani ilość nie mogą być mniejsze niż 1") 
+    return (product, price, quantity, total)
 
 if __name__ == "__main__":      
     main()
